# src/app.ts
import "dotenv/config";
import path from "node:path";
import express, { type Express, type Request, type Response } from "express";
import session from "express-session";
import connectPgSimple from "connect-pg-simple";
import helmet from "helmet";
import rateLimit from "express-rate-limit";
import * as Sentry from "@sentry/node";
import { Pool } from "pg";
import { z } from "zod";
import { prisma } from "./lib/prisma.ts";
import { HttpError } from "./lib/errors.ts";
import { errorHandler } from "./middleware/errorHandler.ts";
import { requireAuth } from "./middleware/requireAuth.ts";
import { getOrCreateDevOwner } from "./lib/currentUser.ts";
import { getLoginRedirectUrl, verifySteamCallback } from "./lib/steamAuth.ts";
import { importOwnedGames } from "./services/steamService.ts";
import { enrichGames } from "./services/igdbService.ts";
import { EstimateSource } from "./lib/estimateSource.ts";
import { createPlan, getPlan, listPlans } from "./services/planService.ts";

const APP_BASE_URL = process.env.APP_BASE_URL ?? "http://localhost:3000";
const FRONTEND_URL = process.env.FRONTEND_URL ?? "http://localhost:5173";

// A real, genuine no-op without SENTRY_DSN set — this repo has no Sentry
// account behind it, so error reporting stays inert until someone adds one.
// init() must run before the routes below so it can actually instrument them.
if (process.env.SENTRY_DSN) {
  Sentry.init({ dsn: process.env.SENTRY_DSN, environment: process.env.NODE_ENV });
}

if (process.env.NODE_ENV === "production" && !process.env.SESSION_SECRET) {
  throw new Error("SESSION_SECRET must be set in production");
}

const createPlanSchema = z.object({ hoursAvailable: z.number().positive() }).strict();
const playSessionSchema = z.object({ hoursPlayed: z.number().positive() }).strict();

// timeToBeatSource is never client-settable directly — it's derived from
// whether timeToBeatHours was just set, cleared, or left alone (see PATCH below).
const gameUpdateSchema = z
  .object({
    name: z.string().min(1).optional(),
    playtimeMinutes: z.number().int().nonnegative().optional(),
    timeToBeatHours: z.number().positive().nullable().optional(),
  })
  .strict()
  .refine((data) => Object.keys(data).length > 0, { message: "At least one field is required" });

export const app: Express = express();

// Real deploys sit behind a TLS-terminating reverse proxy (Render/Fly/Vercel
// all work this way), so Express itself sees a plain HTTP connection — trust
// proxy makes it read "X-Forwarded-Proto" instead, which cookie.secure:"auto"
// below then checks per-request.
if (process.env.NODE_ENV === "production") {
  app.set("trust proxy", 1);
}

// img-src needs the Steam CDN allowed explicitly (helmet's default CSP is
// same-origin only) — everything else stays at helmet's default, which is
// what actually gives us X-Frame-Options, Strict-Transport-Security, etc.
app.use(
  helmet({
    contentSecurityPolicy: {
      directives: {
        ...helmet.contentSecurityPolicy.getDefaultDirectives(),
        "img-src": ["'self'", "data:", "https://cdn.cloudflare.steamstatic.com"],
      },
    },
  }),
);

// Auth endpoints get a generous but real limit — Steam's own callback
// verification already rejects forged logins cryptographically, this is
// defense in depth against brute-forcing that check.
app.use(
  "/auth",
  rateLimit({ windowMs: 15 * 60 * 1000, limit: 20, standardHeaders: true, legacyHeaders: false }),
);

// Import/enrich call real third-party APIs (Steam Web API, IGDB) that are
// themselves rate-limited — this stops a buggy client or an impatient
// double-click from hammering them well before IGDB's own 429s would.
const externalApiLimiter = rateLimit({
  windowMs: 15 * 60 * 1000,
  limit: 5,
  standardHeaders: true,
  legacyHeaders: false,
  message: { error: "Too many import/enrich requests — please wait a few minutes and try again." },
});

// Middleware to parse JSON bodies
app.use(express.json());

// express-session's default MemoryStore leaks memory and doesn't survive a
// restart or scale past one process — fine for local dev, not for a real
// deploy. Postgres is already provisioned, so sessions live there instead;
// createTableIfMissing means no separate migration is needed for it.
const sessionStore =
  process.env.NODE_ENV === "production"
    ? new (connectPgSimple(session))({
        pool: new Pool({ connectionString: process.env.DATABASE_URL }),
        createTableIfMissing: true,
      })
    : undefined;

app.use(
  session({
    store: sessionStore,
    secret: process.env.SESSION_SECRET ?? "dev-secret-change-me",
    resave: false,
    saveUninitialized: false,
    cookie: {
      httpOnly: true,
      sameSite: "lax",
      // "auto" (not "true"): matches the real connection security per-request,
      // so this also works for a plain-HTTP docker-compose smoke test locally.
      secure: process.env.NODE_ENV === "production" ? "auto" : false,
    },
  }),
);

// Health route
app.get('/health', (req: Request, res: Response) => {
  res.json({ status: 'ok' });
});

// --- Steam OpenID login ---

app.get('/auth/steam/login', (req: Request, res: Response) => {
  const returnTo = `${APP_BASE_URL}/auth/steam/callback`;
  res.redirect(getLoginRedirectUrl(returnTo, APP_BASE_URL));
});

app.get('/auth/steam/callback', async (req: Request, res: Response) => {
  const steamId = await verifySteamCallback(req.query as Record<string, string | undefined>);
  if (!steamId) {
    throw new HttpError(401, 'Steam login verification failed');
  }

  const user = await prisma.user.upsert({
    where: { steamId },
    update: {},
    create: { steamId },
  });

  req.session.userId = user.id;
  res.redirect(FRONTEND_URL);
});

app.post('/auth/logout', (req: Request, res: Response) => {
  req.session.destroy(() => {
    res.json({ ok: true });
  });
});

app.get('/auth/me', (req: Request, res: Response) => {
  res.json({ userId: req.session.userId ?? null });
});

// Dev-only convenience: a real session without a real Steam login, since
// that needs a real browser and real Steam credentials neither CI nor an
// agent has. Never mounted in production.
if (process.env.NODE_ENV !== 'production') {
  app.post('/auth/dev-login', async (req: Request, res: Response) => {
    const user = await getOrCreateDevOwner();
    req.session.userId = user.id;
    res.json({ userId: user.id });
  });
}

// --- Everything below operates on the logged-in user's own data ---
app.use('/api', requireAuth);

// List games — ?missing=true limits to games with no time-to-beat estimate yet
app.get('/api/games', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const missing = req.query.missing === 'true';
  const games = await prisma.game.findMany({
    where: { userId, ...(missing ? { timeToBeatHours: null } : {}) },
    orderBy: { id: 'asc' },
  });
  res.json(games);
});

// Manual time-to-beat overrides (and other field updates)
app.patch('/api/games/:id', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const id = Number(req.params.id);
  if (!Number.isInteger(id)) {
    throw new HttpError(400, 'Invalid game id');
  }

  const owned = await prisma.game.findFirst({ where: { id, userId } });
  if (!owned) {
    throw new HttpError(404, 'Not found');
  }

  const { timeToBeatHours, ...rest } = gameUpdateSchema.parse(req.body);

  const data: typeof rest & { timeToBeatHours?: number | null; timeToBeatSource?: EstimateSource } = rest;
  if (timeToBeatHours !== undefined) {
    data.timeToBeatHours = timeToBeatHours;
    data.timeToBeatSource = timeToBeatHours === null ? EstimateSource.NONE : EstimateSource.MANUAL;
  }

  const game = await prisma.game.update({
    where: { id },
    data,
  });
  res.json(game);
});

// Log a play session against a game — advances playtime and lastPlayedAt so
// the next generated plan reflects real progress (smaller remainingHours,
// bigger completionBonus, and the recency penalty kicking in).
app.post('/api/games/:id/play-sessions', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const id = Number(req.params.id);
  if (!Number.isInteger(id)) {
    throw new HttpError(400, 'Invalid game id');
  }

  const owned = await prisma.game.findFirst({ where: { id, userId } });
  if (!owned) {
    throw new HttpError(404, 'Not found');
  }

  const { hoursPlayed } = playSessionSchema.parse(req.body);

  const game = await prisma.game.update({
    where: { id },
    data: {
      playtimeMinutes: { increment: Math.round(hoursPlayed * 60) },
      lastPlayedAt: new Date(),
    },
  });

  res.json(game);
});

// Import owned games from Steam for the logged-in user
app.post('/api/import/steam', externalApiLimiter, async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const result = await importOwnedGames(userId);
  res.json(result);
});

// Enrich games missing a time-to-beat estimate or genre via IGDB
app.post('/api/enrich/igdb', externalApiLimiter, async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const result = await enrichGames(userId);
  res.json(result);
});

// Generate a plan from the current library and persist it (Plan + PlanEntries, atomically)
app.post('/api/plans', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const { hoursAvailable } = createPlanSchema.parse(req.body);
  const plan = await createPlan(userId, hoursAvailable);
  res.status(201).json(plan);
});

// List all plans for the logged-in user, most recent first
app.get('/api/plans', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const plans = await listPlans(userId);
  res.json(plans);
});

// Fetch a single plan with its ordered entries
app.get('/api/plans/:id', async (req: Request, res: Response) => {
  const userId = req.session.userId as number;
  const id = Number(req.params.id);
  if (!Number.isInteger(id)) {
    throw new HttpError(400, 'Invalid plan id');
  }

  const plan = await getPlan(id, userId);
  if (!plan) {
    throw new HttpError(404, 'Not found');
  }

  res.json(plan);
});

// Self-hosted deploys (Docker/Render/Fly) ship the built frontend in the same
// image/process as the API, so the browser talks to one origin and the
// session cookie's sameSite=lax just works — no CORS setup needed. Vercel
// deploys the frontend as static output separately instead (see vercel.json),
// so this block is skipped there. In dev, Vite's own server + proxy plays
// the same role.
if (process.env.NODE_ENV === "production" && !process.env.VERCEL) {
  const frontendDist = path.join(import.meta.dirname, "../frontend/dist");
  app.use(express.static(frontendDist));
  app.use((req: Request, res: Response, next) => {
    if (req.method !== "GET" || req.path.startsWith("/api") || req.path.startsWith("/auth")) {
      next();
      return;
    }
    res.sendFile(path.join(frontendDist, "index.html"));
  });
}

// Unmatched routes
app.use((req: Request, res: Response) => {
  res.status(404).json({ error: 'Not found' });
});

// Reports to Sentry (no-op if SENTRY_DSN was never set) before falling
// through to our own error handler, which still owns the actual response.
if (process.env.SENTRY_DSN) {
  Sentry.setupExpressErrorHandler(app);
}

// Central error handler — must be registered last
app.use(errorHandler);